+++ Connector/TransferObject/Payment/Payment.php
@@ -50,9 +50,14 @@ class Payment extends AbstractTransferObject implements AttributableInterface
     private $transactionReference = '';
 
     /**
-     * @var PaymentDataInterface[]
+     * @var string
      */
-    private $paymentData = [];
+    private $accountHolder = '';
+
+    /**
+     * @var null|PaymentDataInterface
+     */
+    private $paymentData;
 
     /**
      * @var Attribute[]
@@ -184,7 +189,23 @@ class Payment extends AbstractTransferObject implements AttributableInterface
     }
 
     /**
-     * @return PaymentDataInterface[]
+     * @return string
+     */
+    public function getAccountHolder()
+    {
+        return $this->accountHolder;
+    }
+
+    /**
+     * @param string $accountHolder
+     */
+    public function setAccountHolder($accountHolder = '')
+    {
+        $this->accountHolder = $accountHolder;
+    }
+
+    /**
+     * @return null|PaymentDataInterface
      */
     public function getPaymentData()
     {
@@ -192,9 +213,9 @@ class Payment extends AbstractTransferObject implements AttributableInterface
     }
 
     /**
-     * @param PaymentDataInterface $paymentData
+     * @param null|PaymentDataInterface $paymentData
      */
-    public function setPaymentData(PaymentDataInterface $paymentData)
+    public function setPaymentData(PaymentDataInterface $paymentData = null)
     {
         $this->paymentData = $paymentData;
     }